pyproject.toml

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "stuco_app"
version = "1.0.0"
description = "A Web Application created with the Django Framework"
authors = ["Tyler Williams <TylerOgawaWilliams@github.com>"]
repository = "https://github.com/TylerOgawaWilliams/stuco_app"
include = [
    "stuco_app",
    "README.md",
    "pyproject.toml"
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
gunicorn = "^21.2.0"
mozilla-django-oidc = "^4.0.0"
whitenoise = "^6.6.0"
psycopg2-binary = "^2.9.9"
django-htmx = "^1.17.2"
django-auditlog = "^2.3.0"
django-login-history2 = "^0.0.4"
django-widget-tweaks = "^1.5.0"
django-storages = "^1.14.3"
boto3 = "^1.34.142"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.7"
bumpversion = "^0.6.0"
cookiecutter = "^2.5.0"
duty = "^1.2.0"
future = "^0.18.2"
git-changelog = "^2.4.0"
httpx = "^0.26.0"
markupsafe = "2.1.5"
mock = "^5.1.0"
mypy = "^1.8.0"
numpy = "^1.26.4"
pre-commit = "^3.6.0"
pygments = "<2.17.2"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.6.1"
pytest-randomly = "3.5.0"  # pinned to 3.5.0 only on 04-02-2021 as 3.5.1 breaks builds
pytest-sugar = "^1.0.0"
pytest-xdist = "^3.5.0"
pyupgrade = "^3.15.0"
recommonmark = "^0.7.1"
safety = "^3.0.1"
sphinx = "^7.2.6"
sphinx-click = "^5.1.0"
sphinx-rtd-theme = "*"
toml = "^0.10.2"
types-requests = "^2.28.1"
ruff = "^0.2"
django-debug-toolbar = "^4.3.0"

[tool.isort]
# https://github.com/timothycrosley/isort/
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
default_section = "FIRSTPARTY"
multi_line_output = 3
indent = 4
force_grid_wrap = 0
use_parentheses = true
line_length = 100
balanced_wrapping = true

[tool.poetry.scripts]
stuco_app = "stuco_app.cli.cli:cli"

[tool.ruff]
# Enable Pyflakes `E` and `F` codes by default.
lint.select = ["E", "F"]
lint.ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
# lint.fixable = ["A", "B", "C", "D", "E", "F", "..."]
lint.unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".venv",
    ".spg",
    "tests",
]

# Same as Black.
line-length = 120

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"users/migrations/0001_initial.py" = ["E501"]
"stuco_app/cli/cli.py" = ["E501", "F841"]
